added helper methods to Library class
